Move DSID and dsid_to_uuid to model layer.
