+++ datacube/index/abstract/_lineage.py
@@ -20,9 +20,8 @@ if TYPE_CHECKING:
     from uuid import UUID
 
     from datacube.model import LineageDirection, LineageRelation, LineageTree
+    from datacube.model._base import DSID
     from datacube.model.lineage import LineageRelations
-
-    from ._types import DSID
 
 _LOG: logging.Logger = logging.getLogger(__name__)
